pyproject.toml
[tool.poetry]
name = "ChiaFarmerManager"
version = "0.1.0"
description = "A backend with FastAPI, Celery and SQLAlchemy to control flow of Chia plotting and harvesting on multiple machines"
authors = ["LeaveMyYard"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = {extras = ["standart"], version = "^0.13.3"}
celery = "^5.0.5"
SQLAlchemy = "^1.3.23"
python-dotenv = "^0.15.0"
redis = "^3.5.3"
email-validator = "^1.1.2"
mysql-connector = "^2.2.9"
mysql-connector-python = "^8.0.23"
requests = "^2.25.1"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
emails = "^0.6"
python-multipart = "^0.0.5"
Jinja2 = "^2.11.3"
pytest-xdist = "^2.2.1"
boto3 = "^1.17.41"
aiofiles = "^0.6.0"
pandas = "^1.2.3"
websockets = "^8.1"
xlrd = "^2.0.1"
openpyxl = "^3.0.7"
odfpy = "^1.4.1"
fastapi-utils = "^0.2.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.812"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
sqlalchemy-stubs = "^0.4"
rope = "^0.18.0"
pylint = "^2.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.messages_control]
disable = "C0330, C0326, F0401, W0613"

[tool.pylint.format]
max-line-length = "88"